Refactored mscan tests to have abstract parent.
