+++ org.eclipse.scanning.test/src/org/eclipse/scanning/test/command/MScanServletTest.java
@@ -11,170 +11,17 @@
  *******************************************************************************/
 package org.eclipse.scanning.test.command;
 
-import static org.junit.Assert.assertEquals;
-
-import java.net.URI;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.TimeUnit;
-
-import org.eclipse.dawnsci.analysis.api.io.ILoaderService;
-import org.eclipse.dawnsci.hdf5.nexus.NexusFileFactoryHDF5;
-import org.eclipse.dawnsci.json.MarshallerService;
-import org.eclipse.dawnsci.nexus.builder.impl.DefaultNexusBuilderFactory;
-import org.eclipse.dawnsci.remotedataset.test.mock.LoaderServiceMock;
-import org.eclipse.scanning.api.device.IDeviceWatchdogService;
-import org.eclipse.scanning.api.device.IRunnableDeviceService;
-import org.eclipse.scanning.api.device.IScannableDeviceService;
-import org.eclipse.scanning.api.event.EventConstants;
-import org.eclipse.scanning.api.event.IEventService;
-import org.eclipse.scanning.api.event.core.ISubmitter;
-import org.eclipse.scanning.api.event.core.ISubscriber;
-import org.eclipse.scanning.api.event.scan.IScanListener;
-import org.eclipse.scanning.api.event.scan.ScanBean;
-import org.eclipse.scanning.api.event.scan.ScanEvent;
-import org.eclipse.scanning.api.event.status.Status;
-import org.eclipse.scanning.api.points.IPointGeneratorService;
-import org.eclipse.scanning.connector.activemq.ActivemqConnectorService;
-import org.eclipse.scanning.event.EventServiceImpl;
-import org.eclipse.scanning.example.classregistry.ScanningExampleClassRegistry;
-import org.eclipse.scanning.example.detector.MandelbrotDetector;
-import org.eclipse.scanning.example.detector.MandelbrotModel;
-import org.eclipse.scanning.example.malcolm.DummyMalcolmDevice;
-import org.eclipse.scanning.example.malcolm.DummyMalcolmModel;
-import org.eclipse.scanning.example.scannable.MockScannableConnector;
-import org.eclipse.scanning.points.PointGeneratorService;
-import org.eclipse.scanning.points.ScanPointGeneratorFactory;
-import org.eclipse.scanning.points.classregistry.ScanningAPIClassRegistry;
-import org.eclipse.scanning.points.serialization.PointsModelMarshaller;
-import org.eclipse.scanning.points.validation.ValidatorService;
-import org.eclipse.scanning.sequencer.RunnableDeviceServiceImpl;
-import org.eclipse.scanning.sequencer.ServiceHolder;
-import org.eclipse.scanning.sequencer.watchdog.DeviceWatchdogService;
-import org.eclipse.scanning.server.servlet.ScanServlet;
-import org.eclipse.scanning.server.servlet.Services;
-import org.eclipse.scanning.test.ScanningTestClassRegistry;
-import org.eclipse.scanning.test.scan.mock.MockDetectorModel;
-import org.eclipse.scanning.test.scan.mock.MockWritableDetector;
-import org.eclipse.scanning.test.scan.mock.MockWritingMandelbrotDetector;
-import org.eclipse.scanning.test.scan.mock.MockWritingMandlebrotModel;
-import org.junit.AfterClass;
-import org.junit.Before;
-import org.junit.BeforeClass;
 import org.junit.Test;
 
-public class MScanServletTest extends AbstractJythonTest {
+public class MScanServletTest extends AbstractMScanTest {
 
-	
-	private static ScanServlet servlet; 
-	
-	@BeforeClass
-	public static void init() {
-		ScanPointGeneratorFactory.init();
-	}
-	
-	private static IRunnableDeviceService      dservice;
-	private static IScannableDeviceService     connector;
-	private static IPointGeneratorService      gservice;
-	private static IEventService               eservice;
-	private static ILoaderService              lservice;
-	private static IDeviceWatchdogService      wservice;
-	private static MarshallerService           marshaller;
-	private static ValidatorService            validator;
-
-	@BeforeClass
-	public static void create() throws Exception {
-		
-		marshaller = new MarshallerService(
-				Arrays.asList(new ScanningAPIClassRegistry(),
-						new ScanningExampleClassRegistry(),
-						new ScanningTestClassRegistry()),
-				Arrays.asList(new PointsModelMarshaller())
-				);
-		ActivemqConnectorService.setJsonMarshaller(marshaller);
-		eservice  = new EventServiceImpl(new ActivemqConnectorService());
-		
-		// We wire things together without OSGi here
-		// DO NOT COPY THIS IN NON-TEST CODE
-		connector = new MockScannableConnector(null);
-		dservice  = new RunnableDeviceServiceImpl(connector);
-		RunnableDeviceServiceImpl impl = (RunnableDeviceServiceImpl)dservice;
-		impl._register(MockDetectorModel.class, MockWritableDetector.class);
-		impl._register(MockWritingMandlebrotModel.class, MockWritingMandelbrotDetector.class);
-		impl._register(MandelbrotModel.class, MandelbrotDetector.class);
-		impl._register(DummyMalcolmModel.class, DummyMalcolmDevice.class);
-		
-		final MockDetectorModel dmodel = new MockDetectorModel();
-		dmodel.setName("detector");
-		dmodel.setExposureTime(0.1);
-		impl.createRunnableDevice(dmodel);
-
-		MandelbrotModel model = new MandelbrotModel("p", "q");
-		model.setName("mandelbrot");
-		model.setExposureTime(0.00001);
-		impl.createRunnableDevice(model);
-
-		gservice  = new PointGeneratorService();
-		wservice = new DeviceWatchdogService();
-		lservice = new LoaderServiceMock();
-	
-		// Provide lots of services that OSGi would normally.
-		Services.setEventService(eservice);
-		org.eclipse.scanning.command.Services.setEventService(eservice);
-		Services.setRunnableDeviceService(dservice);
-		org.eclipse.scanning.command.Services.setRunnableDeviceService(dservice);
-		Services.setGeneratorService(gservice);
-		Services.setConnector(connector);
-		Services.setWatchdogService(wservice);
-
-		ServiceHolder.setTestServices(lservice, new DefaultNexusBuilderFactory(), null, null, gservice);
-		org.eclipse.scanning.example.Services.setPointGeneratorService(gservice);
-		org.eclipse.dawnsci.nexus.ServiceHolder.setNexusFileFactory(new NexusFileFactoryHDF5());
-		
-		validator = new ValidatorService();
-		validator.setPointGeneratorService(gservice);
-		validator.setRunnableDeviceService(dservice);
-		Services.setValidatorService(validator);
-		
-		
-		// Create an object for the servlet
-		/**
-		 *  This would be done by spring on the GDA Server
-		 *  @see org.eclipse.scanning.server.servlet.AbstractConsumerServlet
-		 *  In spring we have something like:
-
-		    {@literal <bean id="scanner" class="org.eclipse.scanning.server.servlet.ScanServlet">}
-		    {@literal    <property name="broker"      value="tcp://p45-control:61616" />}
-		    {@literal    <property name="submitQueue" value="uk.ac.diamond.p45.submitQueue" />}
-		    {@literal    <property name="statusSet"   value="uk.ac.diamond.p45.statusSet"   />}
-		    {@literal    <property name="statusTopic" value="uk.ac.diamond.p45.statusTopic" />}
-		    {@literal </bean>}
-
-		 */
-		servlet = new ScanServlet();
-		servlet.setBroker(uri.toString());
-		servlet.connect(); // Gets called by Spring automatically
-
+	public MScanServletTest() {
+		super(false);
 	}
 
-	@AfterClass
-	public static void disconnect()  throws Exception {
-		servlet.getConsumer().cleanQueue(servlet.getSubmitQueue());
-		servlet.getConsumer().cleanQueue(servlet.getStatusSet());
-		servlet.disconnect();
-	}
-	
-	@Before
-	public void before() throws Exception {
-		servlet.getConsumer().cleanQueue(servlet.getSubmitQueue());
-		servlet.getConsumer().cleanQueue(servlet.getStatusSet());
-	}
-	
 	@Test
 	public void testGridScan() throws Exception {
-		pi.exec("sr = scan_request(grid(axes=('p', 'q'), start=(0, 0), stop=(10, 10), count=(5, 5), snake=True), det=detector('mandelbrot', 0.1))");
+		pi.exec("sr = scan_request(grid(axes=('xNex', 'yNex'), start=(0, 0), stop=(10, 10), count=(5, 5), snake=True), det=detector('mandelbrot', 0.1))");
 		runAndCheck("sr", false, 10);
 	}
 	
@@ -186,72 +33,14 @@ public class MScanServletTest extends AbstractJythonTest {
 	
 	@Test
 	public void testGridScanNoDetector() throws Exception {
-		pi.exec("sr = scan_request(grid(axes=('p', 'q'), start=(0, 0), stop=(10, 10), count=(5, 5), snake=True))");
+		pi.exec("sr = scan_request(grid(axes=('xNex', 'yNex'), start=(0, 0), stop=(10, 10), count=(5, 5), snake=True))");
 		runAndCheck("sr", false, 10);
 	}
 	
 	@Test
 	public void testGridWithROIScan() throws Exception {
-		pi.exec("sr = scan_request(grid(axes=('p', 'q'), start=(0.0, 1.0), stop=(10.0, 12.0), count=(3, 4), snake=False, roi=[circ(origin=(0.0, 0.0), radius=1.0)]), det=detector('mandelbrot', 0.1))");
+		pi.exec("sr = scan_request(grid(axes=('xNex', 'yNex'), start=(0.0, 1.0), stop=(10.0, 12.0), count=(3, 4), snake=False, roi=[circ(origin=(0.0, 0.0), radius=1.0)]), det=detector('mandelbrot', 0.1))");
 		runAndCheck("sr", false, 10);
-	}
-	
-
-	private List<ScanBean> runAndCheck(String name, boolean blocking, long maxScanTimeS) throws Exception {
-		
-		final IEventService eservice = Services.getEventService();
-
-		// Let's listen to the scan and see if things happen when we run it
-		final ISubscriber<IScanListener> subscriber = eservice.createSubscriber(new URI(servlet.getBroker()), servlet.getStatusTopic());
-		final ISubmitter<ScanBean>       submitter  = eservice.createSubmitter(new URI(servlet.getBroker()),  servlet.getSubmitQueue());
-		
-		try {
-			final List<ScanBean> beans = new ArrayList<>(13);
-			final List<ScanBean> failed = new ArrayList<>(13);
-			final List<ScanBean> startEvents = new ArrayList<>(13);
-			
-			final CountDownLatch latch = new CountDownLatch(1);
-			subscriber.addListener(new IScanListener() {
-				@Override
-				public void scanEventPerformed(ScanEvent evt) {
-					if (evt.getBean().getStatus()==Status.FAILED) failed.add(evt.getBean());
-					if (evt.getBean().getPosition()!=null) {
-						beans.add(evt.getBean());
-					}
-				}
-	
-				@Override
-				public void scanStateChanged(ScanEvent evt) {
-					if (evt.getBean().getStatus()==Status.FAILED) failed.add(evt.getBean());
-					if (evt.getBean().scanStart()) {
-						startEvents.add(evt.getBean()); // Should be just one
-					}
-	                if (evt.getBean().scanEnd()) {
-	                	latch.countDown();
-	                }
-				}
-			});
-	
-			
-			// Ok done that, now we sent it off...
-			pi.exec("submit("+name+", block="+(blocking?"True":"False")+", broker_uri='"+uri+"')");
-			
-			Thread.sleep(200);
-			boolean ok = latch.await(maxScanTimeS, TimeUnit.SECONDS);
-			if (!ok) throw new Exception("The latch broke before the scan finished!");
-			
-			if (failed.size()>0) throw new Exception(failed.get(0).getMessage());
-			
-			ScanBean start = startEvents.get(0);
-			assertEquals(start.getSize(), beans.size());
-			assertEquals(1, startEvents.size());
-			
-			return beans;
-			
-		} finally {
-			subscriber.disconnect();
-			submitter.disconnect();
-		}
 	}
 
 }
